# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Progetto Gruppo T31 - IS2022",
        "description": "API del progetto del Gruppo T31 svolto durente il corso di Ingegneria del software - Anno Accademico 2022/23",
        "contact": {
            "name": "Luca Dematté - Referente gruppo T31",
            "email": "luca.dematte-2@studenti.unitn.it"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "gruppo-t31-api.onrender.com",
    "basePath": "/",
...
    "schemes": ["http", "https"],
    "paths": {
        "/utente/registrazione": {
            "post": {
                "tags": ["utente"],
                "summary": "Creazione di un nuovo account utente",
                "description": "",
                "operationId": "registrazione",
                "consumes": ["multipart/form-data"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "nome",
                        "in": "formData",
                        "description": "Nome del nuovo utente",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "cognome",
                        "in": "formData",
                        "description": "Cognome del nuovo utente",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email del nuovo utente",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password del nuovo utente",
                        "type": "string",
                        "format": "password",
                        "required": true
                    },
                    {
                        "name": "telefono",
                        "in": "formData",
                        "description": "Contatto telefonico del nuovo utente",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "indirizzo",
                        "in": "formData",
                        "description": "Indirizzo del nuovo utente, per l'invio di eventuali comunicazioni scritte",
                        "type": "string",
                        "required": false
                    },
                    {
                        "name": "foto",
                        "in": "formData",
                        "description": "Foto profilo del nuovo utente, quadrata ed in formato .png",
                        "type": "file",
                        "required": false
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created - Registrazione effettuata correttamente",
                        "schema": { "$ref": "#/definitions/Utente" }
                    },
                    "400": {
                        "description": "Bad request - I parametri richiesti non sono specificati",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "409": {
                        "description": "Conflict - E' già presente un utente con l'email specificata",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
...
        "/utente/conferma": {
            "get": {
                "tags": ["utente"],
                "summary": "Conferma registrazione dell'account",
                "description": "",
                "operationId": "confermaAccount",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID dell'utente che sta confermando la registrazione",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer",
                                    "description": "Codice di risposta HTTP",
                                    "example": 200
                                },
                                "message": {
                                    "type": "string",
                                    "description": "Messaggio con esito dell'operazione",
                                    "example": "L'utente con l'id <ID utente> è stato confermato"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - ID non specificato o non valido",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Un utente con l'ID fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
        "/utente": {
            "get": {
                "tags": ["utente"],
                "summary": "Ottenere la lista degli utenti presenti nel sistema",
                "description": "",
                "operationId": "listaUtenti",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Indice del primo oggetto da restituire",
                        "type": "integer",
                        "required": false
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Numero di oggetti da restituire",
                        "type": "integer",
                        "required": false
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": { "$ref": "#/definitions/Utente" }
                        }
                    },
                    "204": {
                        "description": "No content - Nessun utente da restituire",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer",
                                    "description": "Codice di risposta HTTP"
                                },
                                "message": {
                                    "type": "string",
                                    "description": "Messaggio con esito dell'operazione"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - I valori di start e count non sono coerenti",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
        "/utente/byID": {
            "get": {
                "tags": ["utente"],
                "summary": "Ottenere un utente tramite il suo ID",
                "description": "",
                "operationId": "getUtenteConID",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID dell'utente da restituire",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": { "$ref": "#/definitions/Utente" }
                    },
                    "400": {
                        "description": "Bad request - ID non specificato",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Un utente con l'ID fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
        "/utente/byEmail": {
            "get": {
                "tags": ["utente"],
                "summary": "Ottenere un utente tramite la sua email",
                "description": "",
                "operationId": "getUtenteConEmail",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email dell'utente da restituire",
                        "type": "string",
                        "required": true
                    },
                    {
                        "name": "x-access-token",
                        "in": "header",
                        "description": "Token di accesso dell'utente",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": { "$ref": "#/definitions/Utente" }
                    },
                    "400": {
                        "description": "Bad request - Email non specificata",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Un utente con l'email fornita non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
        "/utente/byToken": {
            "get": {
                "tags": ["utente"],
                "summary": "Ottenere un utente tramite un token di accesso",
                "description": "",
                "operationId": "getUtenteConToken",
                "consumes": [],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token legato all'utente da restituire",
                        "type": "string",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": { "$ref": "#/definitions/Utente" }
                    },
                    "400": {
                        "description": "Bad request - Token non specificato",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "401": {
                        "description": "Unauthorized - Per svolgere l'operazione è necessario il login",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "403": {
                        "description": "Forbidden - Il livello di privilegio non è sufficiente",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "404": {
                        "description": "Not found - Un utente legato al token fornito non esiste",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    },
                    "500": {
                        "description": "Internal server error - Errore imprevisto durante l'operazione",
                        "schema": { "$ref": "#/definitions/ErroreAPI" }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Utente": {
            "type": "object",
            "required": ["_id", "nome", "cognome", "email", "confermaAccount", "livello"],
            "properties": {
                "_id": {
                    "type": "string",
                    "description": "Identificativo univoco dell'utente"
                },
                "nome": {
                    "type": "string",
                    "description": "Nome dell'utente",
                    "example": "Mario"
                },
                "cognome": {
                    "type": "string",
                    "description": "Cognome dell'utente",
                    "example": "Rossi"
                },
                "email": {
                    "type": "string",
                    "description": "Email utilizzata per la creazione dell'account",
                    "example": "mario.rossi@mail.com"
                },
                "password": {
                    "type": "string",
                    "description": "Password dell'utente",
                    "example": "<password>"
                },
                "telefono": {
                    "type": "string",
                    "description": "Contatto telefonico dell'utente",
                    "example": "1234567890"
                },
                "indirizzo": {
                    "type": "string",
                    "description": "Indirizzo dell'utente, per l'invio di eventuali comunicazioni scritte",
                    "example": "via Roma 34, 20122 Milano"
                },
                "URLfoto": {
                    "type": "string",
                    "description": "Percorso della foto profilo dell'utente",
                    "example": "/images/utenti/default.png"
                },
                "confermaAccount": {
                    "type": "boolean",
                    "description": "Status dell'account dell'utente, se false allora l'utente deve ancora confermare la registrazione",
                    "example": true
                },
                "livello": {
                    "type": "integer",
                    "description": "Livello di privilegio dell'utente: 0 -> non registrato, 1 -> utente standard, 2 -> utente di segreteria",
                    "example": 1
                }
            }
        },
...
        "ErroreAPI": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
                "code": {
                    "type": "integer",
                    "description": "Codice di risposta HTTP"
                },
                "message": {
                    "type": "string",
                    "description": "Messaggio di testo sull'errore"
                }
            }
        }
    }
}